fixup! Add check when trying to reroute a shard to a non-data discovery node
